make debug functions part of core only, so devs can use their own local functions
