+++ composer.json
@@ -48,7 +48,6 @@
         "symplify/skipper": "^11.1",
         "symplify/smart-file-system": "^11.1",
         "symplify/symplify-kernel": "^11.1",
-        "tracy/tracy": "^2.9.4",
         "webmozart/assert": "^1.11"
     },
     "require-dev": {
@@ -64,6 +63,7 @@
         "phpunit/phpunit": "^9.5.21",
         "rector/phpstan-rules": "^0.5.15",
         "spatie/enum": "^3.13",
+        "tracy/tracy": "^2.9.4",
         "symfony/process": "^6.1",
         "symplify/easy-ci": "^11.1",
         "symplify/easy-coding-standard": "^11.1",
@@ -88,7 +88,6 @@
             "Rector\\Utils\\": "utils"
         },
         "files": [
-            "src/functions/node_helper.php",
             "src/constants.php"
         ]
     },
@@ -112,6 +111,7 @@
             "rules-tests/Renaming/Rector/FileWithoutNamespace/PseudoNamespaceToNamespaceRector/Source"
         ],
         "files": [
+            "tests/functions/debug_functions.php",
             "stubs/Doctrine/Persistence/ObjectManager.php",
             "stubs/Doctrine/Common/Persistence/ObjectManager.php",
             "rules-tests/Transform/Rector/FuncCall/FuncCallToMethodCallRector/Source/some_view_function.php",
