+++ scoper.php
@@ -67,7 +67,7 @@ return [
         'Symplify\SmartFileSystem\SmartFileInfo',
         'Symplify\ComposerJsonManipulator\ValueObject\ComposerJson',
     ],
-    'expose-functions' => ['u', 'b', 's', 'trigger_deprecation', 'dump_with_depth', 'dn', 'dump_node', 'print_node'],
+    'expose-functions' => ['u', 'b', 's', 'trigger_deprecation'],
     'expose-constants' => ['__RECTOR_RUNNING__', '#^SYMFONY\_[\p{L}_]+$#'],
 
     'patchers' => [
