+++ src/PhpParser/Node/Value/ValueResolver.php
@@ -30,6 +30,7 @@ use Rector\NodeTypeResolver\NodeTypeResolver;
 
 /**
  * @see \Rector\Core\Tests\PhpParser\Node\Value\ValueResolverTest
+ * @todo make use of constant type of $scope->getType()
  */
 final class ValueResolver
 {
